Refactor of the commiting and revealing answers
- natspec update

+++ Quiz-Project/test/RewardFunctions.t.sol
@@ -5,24 +5,21 @@ import "../src/Quiz.sol";
 
 contract QuizTest is Test {
     Quiz public quizContract;
-    event Log(string func, uint gas);
+
+    event Log(string func, uint256 gas);
 
     function setUp() public {
         quizContract = new Quiz();
     }
-
-    
 
     function test_RevertWhen_NotWinner() public {
         vm.expectRevert("You do not qualify for a reward");
         quizContract.withdrawReward();
     }
 
-
     function testContractEthBalance() public {
         console.log("ETH Balance", address(quizContract).balance / 1e18);
     }
-
 
     function test_RewardPayment() public {
         vm.deal(address(quizContract), 10 ether);
@@ -30,11 +27,7 @@ contract QuizTest is Test {
         quizContract.withdrawReward();
     }
 
-
     receive() external payable {
         emit Log("receive", gasleft());
     }
-
-    
-
 }
